LPS-59899 Move RSS logic to MVCResourceCommand (no logic changes)
