+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/portlet/JournalPortlet.java
@@ -62,7 +62,6 @@ import com.liferay.journal.util.impl.JournalUtil;
 import com.liferay.journal.web.asset.JournalArticleAssetRenderer;
 import com.liferay.journal.web.configuration.JournalWebConfiguration;
 import com.liferay.journal.web.portlet.action.ActionUtil;
-import com.liferay.journal.web.util.JournalRSSUtil;
 import com.liferay.portal.LocaleException;
 import com.liferay.portal.kernel.diff.CompareVersionsException;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -71,13 +70,11 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.LiferayWindowState;
 import com.liferay.portal.kernel.portlet.PortletRequestModel;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.portal.kernel.servlet.ServletResponseUtil;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.servlet.SessionMessages;
 import com.liferay.portal.kernel.upload.LiferayFileItemException;
 import com.liferay.portal.kernel.upload.UploadException;
 import com.liferay.portal.kernel.upload.UploadPortletRequest;
-import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.HttpUtil;
@@ -488,22 +485,6 @@ public class JournalPortlet extends MVCPortlet {
 					"/compare_versions_diff_html.jsp");
 
 			portletRequestDispatcher.include(resourceRequest, resourceResponse);
-		}
-		else if (resourceID.equals("rss")) {
-			try {
-				byte[] xml = _journalRSSUtil.getRSS(
-					resourceRequest, resourceResponse);
-
-				ServletResponseUtil.sendFile(
-					request, response, null, xml, ContentTypes.TEXT_XML_UTF8);
-			}
-			catch (Exception e) {
-				try {
-					PortalUtil.sendError(e, request, response);
-				}
-				catch (ServletException se) {
-				}
-			}
 		}
 		else {
 			super.serveResource(resourceRequest, resourceResponse);
@@ -1325,11 +1306,6 @@ public class JournalPortlet extends MVCPortlet {
 	}
 
 	@Reference
-	protected void setJournalRSSUtil(JournalRSSUtil journalRSSUtil) {
-		_journalRSSUtil = journalRSSUtil;
-	}
-
-	@Reference
 	protected void setLayoutLocalService(
 		LayoutLocalService layoutLocalService) {
 
@@ -1386,10 +1362,6 @@ public class JournalPortlet extends MVCPortlet {
 		_journalFolderService = null;
 	}
 
-	protected void unsetJournalRSSUtil(JournalRSSUtil journalRSSUtil) {
-		_journalRSSUtil = null;
-	}
-
 	protected void unsetLayoutLocalService(
 		LayoutLocalService layoutLocalService) {
 
@@ -1431,7 +1403,6 @@ public class JournalPortlet extends MVCPortlet {
 	private JournalConverter _journalConverter;
 	private JournalFeedService _journalFeedService;
 	private JournalFolderService _journalFolderService;
-	private JournalRSSUtil _journalRSSUtil;
 	private volatile JournalWebConfiguration _journalWebConfiguration;
 	private LayoutLocalService _layoutLocalService;
 	private TrashEntryService _trashEntryService;
